+++ src/StorageBase/Action/IdentityRedirect/Create/IdentityRedirectCreatePayload.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Heptacom\HeptaConnect\Storage\Base\Action\IdentityDirection\Create;
+namespace Heptacom\HeptaConnect\Storage\Base\Action\IdentityRedirect\Create;
 
 use Heptacom\HeptaConnect\Dataset\Base\AttachmentCollection;
 use Heptacom\HeptaConnect\Dataset\Base\Contract\AttachmentAwareInterface;
@@ -11,7 +11,7 @@ use Heptacom\HeptaConnect\Dataset\Base\Support\AttachmentAwareTrait;
 use Heptacom\HeptaConnect\Portal\Base\StorageKey\Contract\PortalNodeKeyInterface;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\Create\CreatePayloadInterface;
 
-final class IdentityDirectionCreatePayload implements CreatePayloadInterface, AttachmentAwareInterface
+final class IdentityRedirectCreatePayload implements CreatePayloadInterface, AttachmentAwareInterface
 {
     use AttachmentAwareTrait;