+++ src/StorageBase/Action/IdentityRedirect/Create/IdentityRedirectCreatePayloadCollection.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Heptacom\HeptaConnect\Storage\Base\Action\IdentityDirection\Create;
+namespace Heptacom\HeptaConnect\Storage\Base\Action\IdentityRedirect\Create;
 
 use Heptacom\HeptaConnect\Dataset\Base\AttachmentCollection;
 use Heptacom\HeptaConnect\Dataset\Base\Contract\AttachmentAwareInterface;
@@ -10,9 +10,9 @@ use Heptacom\HeptaConnect\Dataset\Base\Support\AbstractObjectCollection;
 use Heptacom\HeptaConnect\Dataset\Base\Support\AttachmentAwareTrait;
 
 /**
- * @extends AbstractObjectCollection<IdentityDirectionCreateResult>
+ * @extends AbstractObjectCollection<IdentityRedirectCreatePayload>
  */
-final class IdentityDirectionCreateResultCollection extends AbstractObjectCollection implements AttachmentAwareInterface
+final class IdentityRedirectCreatePayloadCollection extends AbstractObjectCollection implements AttachmentAwareInterface
 {
     use AttachmentAwareTrait;
 
@@ -22,11 +22,8 @@ final class IdentityDirectionCreateResultCollection extends AbstractObjectCollec
         $this->attachments = new AttachmentCollection();
     }
 
-    /**
-     * @psalm-return IdentityDirectionCreateResult::class
-     */
     protected function getT(): string
     {
-        return IdentityDirectionCreateResult::class;
+        return IdentityRedirectCreatePayload::class;
     }
 }