+++ src/StorageBase/Bridge/Contract/StorageFacadeInterface.php
@@ -10,10 +10,10 @@ use Heptacom\HeptaConnect\Storage\Base\Contract\Action\Identity\IdentityMapActio
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\Identity\IdentityOverviewActionInterface;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\Identity\IdentityPersistActionInterface;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\Identity\IdentityReflectActionInterface;
-use Heptacom\HeptaConnect\Storage\Base\Contract\Action\IdentityDirection\IdentityDirectionCreateActionInterface;
-use Heptacom\HeptaConnect\Storage\Base\Contract\Action\IdentityDirection\IdentityDirectionDeleteActionInterface;
-use Heptacom\HeptaConnect\Storage\Base\Contract\Action\IdentityDirection\IdentityDirectionOverviewActionInterface;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\IdentityError\IdentityErrorCreateActionInterface;
+use Heptacom\HeptaConnect\Storage\Base\Contract\Action\IdentityRedirect\IdentityRedirectCreateActionInterface;
+use Heptacom\HeptaConnect\Storage\Base\Contract\Action\IdentityRedirect\IdentityRedirectDeleteActionInterface;
+use Heptacom\HeptaConnect\Storage\Base\Contract\Action\IdentityRedirect\IdentityRedirectOverviewActionInterface;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\Job\JobCreateActionInterface;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\Job\JobDeleteActionInterface;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\Job\JobFailActionInterface;
@@ -73,25 +73,25 @@ interface StorageFacadeInterface
     public function getFileReferencePersistRequestAction(): FileReferencePersistRequestActionInterface;
 
     /**
-     * Provides storage action to create directional identities.
+     * Provides storage action to create identity redirects.
      *
      * @throws StorageFacadeServiceExceptionInterface
      */
-    public function getIdentityDirectionCreateAction(): IdentityDirectionCreateActionInterface;
+    public function getIdentityRedirectCreateAction(): IdentityRedirectCreateActionInterface;
 
     /**
-     * Provides storage action to delete directional identities.
+     * Provides storage action to delete identity redirects.
      *
      * @throws StorageFacadeServiceExceptionInterface
      */
-    public function getIdentityDirectionDeleteAction(): IdentityDirectionDeleteActionInterface;
+    public function getIdentityRedirectDeleteAction(): IdentityRedirectDeleteActionInterface;
 
     /**
-     * Provides storage action to paginate over all directional identities.
+     * Provides storage action to paginate over all identity redirects.
      *
      * @throws StorageFacadeServiceExceptionInterface
      */
-    public function getIdentityDirectionOverviewAction(): IdentityDirectionOverviewActionInterface;
+    public function getIdentityRedirectOverviewAction(): IdentityRedirectOverviewActionInterface;
 
     /**
      * Provides storage action to map entities to their identities.