+++ src/TestSuiteStorage/CHANGELOG.md
@@ -9,8 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add `\Heptacom\HeptaConnect\TestSuite\Storage\Action\IdentityMappingTestContract::testReflectTwoEntitiesOfSameTypeFromPortalNodeAToBWithDirectionalMappings` to test identities using directional identities
-- Add `\Heptacom\HeptaConnect\TestSuite\Storage\Action\IdentityMappingTestContract::testReflectTwoEntitiesOfSameTypeFromPortalNodeAToBWithPortalAssignedAndDirectionalMappings` to test identities using directional identities and portal assigned identities
+- Add `\Heptacom\HeptaConnect\TestSuite\Storage\Action\IdentityMappingTestContract::testReflectTwoEntitiesOfSameTypeFromPortalNodeAToBWithIdentityRedirects` to test identities using identity redirects
+- Add `\Heptacom\HeptaConnect\TestSuite\Storage\Action\IdentityMappingTestContract::testReflectTwoEntitiesOfSameTypeFromPortalNodeAToBWithPortalAssignedAndIdentityRedirects` to test identities using identity redirects and portal assigned identities
 
 ### Changed